requirements.txt
Flask
flask-cors
mysql-connector-python
translate
gtts
python-dotenv